fix: bundle runtime modules in wheel + auto-release with attestations
Fixes #19: published wheel had `from tools import install` in
clang_tool_chain_bins/__init__.py, but `tools/` was never bundled in
the wheel, so `import clang_tool_chain_bins` raised ModuleNotFoundError.

Refactor (issue's preferred Option 1):
- Move runtime modules (install, query, archive_index, common,
  download_sources, json_utils, cli) into clang_tool_chain_bins/_impl/
- Bundle _impl/data/*.json (tool-index + index-meta) via maturin include
- Replace tools/<runtime>.py with sys.modules backward-compat shims so
  tests + maintainer scripts keep working
- Update entry points: clang_tool_chain_bins._impl.cli:main, etc.
- Bump version 0.4.1 -> 0.4.2 (workspace + inter-crate deps + pyproject)

Bonus from issue:
- Regenerate tool-index.json with linux x86_64 emscripten archives
  (4.0.15, 4.0.21). build_aggregate_index now mirrors the index to
  tools/data/ for backward-compat with older clients fetching the
  raw.githubusercontent URL.

Auto-release (zccache-style):
- release.yml now triggers on push-to-main with version bump, tag push,
  or workflow_dispatch (with skip-if-published guards)
- Adds crates.io publish step (uses CARGO_REGISTRY_TOKEN secret)
- PyPI stays on trusted publishing

Attestations:
- PyPI: PEP 740 attestations via pypa/gh-action-pypi-publish
- crates.io: sigstore provenance via actions/attest-build-provenance
  on the .crate files produced by cargo package
