+++ clang_tool_chain_bins/_impl/__init__.py
@@ -1,0 +1,7 @@
+"""Runtime implementation for clang_tool_chain_bins.
+
+This package contains the install/query/CLI logic that backs the public
+`clang_tool_chain_bins` API. It is bundled inside the wheel; the parallel
+`tools/` directory at the repository root contains maintainer-only build
+scripts that re-export from here for backward compatibility.
+"""
