+++ crates/ctcb-archive/Cargo.toml
@@ -15,7 +15,7 @@ walkdir = { workspace = true }
 anyhow = { workspace = true }
 indicatif = { workspace = true }
 flate2 = { workspace = true }
-ctcb-core = { path = "../ctcb-core", version = "0.4.1" }
+ctcb-core = { path = "../ctcb-core", version = "0.4.2" }
 
 [dev-dependencies]
 tempfile = { workspace = true }
