+++ crates/ctcb-strip/Cargo.toml
@@ -11,7 +11,7 @@ description = "LLVM binary stripping and filtering for clang toolchain binary bu
 [dependencies]
 walkdir = { workspace = true }
 anyhow = { workspace = true }
-ctcb-core = { path = "../ctcb-core", version = "0.4.1" }
+ctcb-core = { path = "../ctcb-core", version = "0.4.2" }
 
 [dev-dependencies]
 tempfile = { workspace = true }
